Added in functionality for endOfGame; adjusted a few parameters
endOfGame is triggered when there are a specified number of null
offers, or when there are no executed trades for a specified number of
rounds.  These are two separate parameters, and we can vary them.

Also, in eatNow(), I made an important change.  If it's end of game, we
still won't eat a pile if the pile is negative.  These skittles are
eaten one-by-one (only the positive piles are eaten in one turn).  This
greatly improved out score.  We now never receive a negative overall
score.